hack to use long double instead of float. Haven't checked to see what effect that has on an actual iPhone, but at very high zooms it makes things look *MUCH* better

+++ MandelbrotSet/RTMandelbrotOperation.h
@@ -27,6 +27,6 @@
 
 - (id)initWithBounds:(CGRect)newBounds retina:(BOOL)ret;
 
-- (float)scaleX:(CGFloat)screenCoord;
-- (float)scaleY:(CGFloat)screenCoord;
+- (long double)scaleX:(CGFloat)screenCoord;
+- (long double)scaleY:(CGFloat)screenCoord;
 @end
